% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorTools.R
\name{colorCroquis}
\alias{colorCroquis}
\title{Colors for pot designs for G line}
\usage{
colorCroquis(line = "G", alphaG = 0.5)
}
\arguments{
\item{line}{: potline default to G line}

\item{alphaG}{: alpha for graphite blocks}
}
\value{
named vector of color for pot designs
}
\description{
Colors for pot designs for G line
}
