% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorTools.R
\name{colorRangeOneColor}
\alias{colorRangeOneColor}
\title{colorRangeOneColor}
\usage{
colorRangeOneColor(
  nColor = 5,
  baseColor = extractColors("greyRed"),
  begin = 0,
  end = 100,
  rev = F
)
}
\arguments{
\item{nColor}{number of colors}

\item{baseColor}{Base color}

\item{begin}{lightness value for start color}

\item{end}{lightness value for end color}

\item{rev}{if TRUE reverse the colors}
}
\value{
vector of colors in hex format
}
\description{
Create a range of colors from one color with different levels of lightness
50 is the brightness of the base color
}
\examples{
shades::swatch(colorRangeOneColor(10))

}
