% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colorVectorPotline}
\alias{colorVectorPotline}
\title{Potline Colors}
\format{
A named vector containing 10 colors
\describe{
...
}
}
\usage{
colorVectorPotline
}
\description{
A vector containing named colors from STJN potline - See getcolors vignette for
details
}
\keyword{datasets}
