% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colorVectorTrimet}
\alias{colorVectorTrimet}
\title{Trimet Colors}
\format{
A named vector containing 2 colors white and blue
\describe{
...
}
}
\usage{
colorVectorTrimet
}
\description{
A vector containing named colors from Trimet Logo - See getcolors vignette for
details
}
\keyword{datasets}
