% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorTools.R
\name{interpolatePalette}
\alias{interpolatePalette}
\title{interpolatePalette}
\usage{
interpolatePalette(
  paletteList = customColors::potlinePalettes,
  paletteName = "redGrey",
  reverse = F,
  ...
)
}
\arguments{
\item{paletteList}{list of named color vectors. Default to customColors::potlinePalettes}

\item{paletteName}{name of palette in palette list.  Default to 'redGrey'}

\item{reverse}{if TRUE reverse he colors}

\item{...}{passed on to colorRampPalette.}
}
\value{
function
}
\description{
Create a function to generate set of n colors from a color palette.
}
\examples{
interpolatePalette()(10)

}
