% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{potlinePalettes}
\alias{potlinePalettes}
\title{Potline Palettes}
\format{
A list containing names vectors with colors from colorVectorPotline
\describe{
...
}
}
\usage{
potlinePalettes
}
\description{
A list of sets (palettes) of colors from colorVectorPotline
}
\keyword{datasets}
