% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorTools.R
\name{scale_color_trimet}
\alias{scale_color_trimet}
\title{scale_color_trimet}
\usage{
scale_color_trimet(palette = "main", discrete = T, reverse = F, ...)
}
\arguments{
\item{palette}{name of palette in customColors::trimetPalette}

\item{discrete}{TRUE default for discrete scale}

\item{reverse}{TRUE to reverse color order}

\item{...}{passed on to discrete_scale or scale_color_gradientn}
}
\value{
scale
}
\description{
Create scales for ggplot using Trimet colors
}
