% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorTools.R
\name{scale_fill_potline}
\alias{scale_fill_potline}
\title{scale_fill_potline}
\usage{
scale_fill_potline(palette = "redBeigeGrey", discrete = T, reverse = F, ...)
}
\arguments{
\item{palette}{name of palette in customColors::potlinePalettes}

\item{discrete}{TRUE default for discrete scale}

\item{reverse}{TRUE to reverse color order}

\item{...}{passed on to discrete_scale or scale_color_gradientn}
}
\value{
scale
}
\description{
Create scales for ggplot using potline colors
}
