% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trimetPalette}
\alias{trimetPalette}
\title{Trimet Palette}
\format{
A list containing names vectors with colors from colorVectorTrimet
\describe{
...
}
}
\usage{
trimetPalette
}
\description{
A list of sets (palettes) of colors from colorVectorTrimet
}
\keyword{datasets}
